backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
nest-asyncio
python-dotenv
pytz
tzlocal
python-telegram-bot